extract <IdentityOrAddr>
