+++ packages/next/services/nextApi.js
@@ -9,7 +9,7 @@ export default new Api(
 
 export const ssrNextApi = new Api(
   new URL(
-    `/`,
+    `/api/`,
     process.env.NEXT_PUBLIC_BACKEND_API_END_POINT ||
       "https://test.opensquare.io/"
   ).href
